requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
configparser==4.0.2
contextlib2==0.6.0.post1
enum34==1.1.10
flake8==3.9.2
functools32==3.2.3.post2
importlib-metadata==2.1.3
kombu==4.6.11
mccabe==0.6.1
pathlib==1.0.1
pathlib2==2.3.7.post1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.18.0
pytz==2022.6
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
vine==1.3.0
zipp==1.2.0
